(amend included after original message)IMPORTANT: An issue on the end of TD's API has been found for its Options Chain API where at least the range (but potentially more) API parameters, even directly on TD's API 'try it out' pages, doesn't do/filter what it says that parameter will do, so options code will need to be reworked and refactored so that ITM/OTM status as well as other search filters can work correctly
(amend here:)
After further investigation, the API requires certain fields with default values in order for other fields to work as intended, so combining most or all of the filters (excluding from/to and exp month at the same time) will result in unintended ignored functionality.  This still needs to be fixed but is different than as originally described.